Move analytics to util and update on url change
